feat(gantt): implement milestone and task components with context menus and drag-resize functionality
- Add MilestoneRow component to display milestones with expandable details and draggable resize handle.
- Create ProjectInfo component to summarize project timeline and milestones.
- Introduce TaskBar component for individual task representation with editing capabilities.
- Develop TaskRow component to manage task display, including context menus for actions like clone, split, move, and delete.
- Implement TimelineHeader for displaying date columns and deliverable markers.
- Add ZoomControls for adjusting the zoom level of the timeline and managing milestone visibility.
